shipsum_t 16->32

+++ src/game/game_types.h
@@ -217,7 +217,7 @@ typedef enum {
 #define TECHSOURCE_CHOOSE   TECHSOURCE_AI_SPY
 
 typedef uint16_t shipcount_t;
-typedef uint16_t shipsum_t;
+typedef uint32_t shipsum_t;
 
 #define SHIP_NAME_NUM   12
